requirements.txt
python-telegram-bot==11.1.0
google-cloud-firestore==1.4.0
argon2-cffi==19.2.0
quantumrandom==1.9.0